+++ backend/lib/messaging/sms/sms-sending-tool.ts
@@ -1,9 +1,9 @@
 import { ArgumentParser } from "argparse"
-import config from "../config/index.js"
-import { SmsCategory } from "../../lib/enums/sms.js"
+import config from "../../../config/index.js"
+import { SmsCategory } from "../../../../lib/enums/sms.js"
 import mongoose from "mongoose"
-import mongooseConfig from "../config/mongoose.js"
-import { processSendSms } from "./sms/sending.js"
+import mongooseConfig from "../../../config/mongoose.js"
+import { processSendSms } from "../../sms/sending.js"
 
 mongooseConfig(mongoose, config)
