+++ backend/lib/messaging/sms/sms-service.ts
@@ -1,5 +1,5 @@
 import axios from "axios"
-import config from "../config/index.js"
+import config from "../../../config/index.js"
 
 export async function renderAndSendSimulationResultsSms(followup) {
   try {
